Add bug checking
If the project is configured with a bugzilla_product and a
bugzilla_component and the user doesn't pass a --with-bug, then that's
probably a user error release.py should point out. This fixes that.

If the user provides a --with-bug, then release.py can download the JSON
for that bug and show the user the summary making it less likely the
user mistypes the bug number. This fixes that, too.