Print url to tag information GitHub (#45)
This changes the end of `make-tag` to print just the url to the tag
information on GitHub rather than print all the tag information. There
will be less to copy-and-paste into Bugzilla plus Bugzilla will now have
a link to the tag information on GitHub which is probably more useful in
the long run.